requirements.txt
django>=4.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
google-genai>=0.3.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
django-tailwind==3.8.0
crispy-tailwind==0.5.0